requirements.txt
certifi==2016.2.28
chardet==2.3.0
contextlib2==0.5.4
coverage==3.7.1
Django==1.8.16
django-anymail==0.6.1
django-braces==1.9.0
django-coverage==1.2.4
django-crispy-forms==1.6.1
django-debug-toolbar==1.5
django-environ==0.4.0
django-extensions==1.6.7
django-floppyforms==1.7.0
django-jsonview==1.0.0
django-redis==4.6.0
django-reversion==1.8.5
django-secure==1.0.1
django-storages==1.5.1
djangorestframework==3.4.0
eventlet==0.19.0
feedparser==5.2.1
github3.py==0.9.6
greenlet==0.4.10
httplib2==0.9.2
linaro-django-pagination==2.0.3
logutils==0.3.3
mimeparse==0.1.3
newrelic==2.74.0.54
oauth2==1.9.0.post1
oauthlib==2.0.2
opbeat==3.5.2
psycopg2==2.6.2
PyJWT==1.4.2
python-dateutil==2.6.0
python-openid==2.2.5
python-social-auth==0.2.3
raven==5.32.0
redis==2.10.5
requests==2.11.1
requests-oauthlib==0.8.0
shortuuid==0.4.3
six==1.10.0
sqlparse==0.2.3
uritemplate==3.0.0
uritemplate.py==3.0.2
urllib3==1.20
uWSGI==2.0.14
webstack-django-sorting==0.4.3
wheel==0.24.0
whitenoise==3.2.2
